requirements.txt
# ── Thesis pipeline (existing) ──────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
nltk>=3.9
scikit-learn>=1.4.0
scipy>=1.11.0
matplotlib>=3.8.0

# ── ReportRx web app ───────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.6.0